\name{lectometry-package}
\alias{lectometry-package}
\alias{lectometry}
\docType{package}
\title{
An R implementation of lectometric methodology
}
\description{
This package implements a number of lectometric methods so that they can easily be used and applied.
}
\details{
\tabular{ll}{
Package: \tab lectometry\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2012-08-28\cr
License: \tab What license is it under?\cr
}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
Tom Ruette

Maintainer: tom.ruette@gmail.com
}
\references{
D. Geeraerts, S. Grondelaers and D. Speelman (1999) Convergentie en divergentie in de Nederlandse woordenschat. Meertens Instituut.
D. Speelman, S. Grondelaers and D. Geeraerts (2003) Profile-based linguistic uniformity as a generic method for comparing language varieties
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
